Performance tweak

+++ Assets/Scripts/Misc/FPSCounter.cs
@@ -9,6 +9,6 @@ public class FPSCounter : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        textField.text = "" + 1f / Time.unscaledDeltaTime;
+        textField.text = "" + (int)(1f / Time.unscaledDeltaTime);
     }
 }
